Back to changing SR2 implementation.
There seems to be a bug with the SR2 buffer decryption. Garbage values for Blacklevel and Whitebalance come out. Usage of SR2 data is currently deactivated.

Further work on a nice GTK frontend is currently started. The goal is to not require changes in source code when common testing parameters need to be changed.