Much improved GUI support.
The GLADE generated UI has some issues. GLADE bugs out when an image is inside of a GTKScrolledWindow. This is added in my hand in to the UI file currently. GLADE also fails to set up a proper popover menu contents wise. This is done from Go code currently, this will also move to patching the UI file in the future.

Dimensions for the image should be taken from the UI file, currently 1500x1000 is used as the baseline in code and 900x600 in the UI file. This leads to inconsistent zoom levels.

Performance of scaling is quite poor at the moment. Every single scaling action starts from zero again from the entire (42MP in some cases) backbuffer. For scaling actions where less than the entire image has to be shown there might be a way to only render the scaled area immediately on display. Because we are in a scrolling window there does have to be an option to get the rest of the image when one scrolls in to not yet rendered space.