Added working metadata display.
Squashed a bug where instead of an & for masking out certain values from the Offset field in an IFD we used *. This caused widly incorrect results.
Spotted a bug with relation to null terminating characters in the strings we read in from IFD ASCII fields. We shouldn't put the null terminator at the end in Go strings. The null terminator causes issues when printing the strings. For now we trim the final character of the one case where this popped up, it will need to be changed in the ASCII handling part of the IFD code.